added avg chart & build
