+++ src/app/cmps/app-header/app-header.component.scss
@@ -23,7 +23,7 @@ header {
   nav {
     display: flex;
     align-items: center;
-    gap: 117px;
+    gap: 25px;
 
     span{
       cursor: pointer;
