+++ src/app/cmps/avg-months-chart/avg-months-chart.component.html
@@ -1,0 +1,5 @@
+<section class="main-container">
+    <div id="divChart">
+      <canvas id="myChart" width="800"></canvas>
+    </div>
+  </section>
