+++ src/app/cmps/avg-months-chart/avg-months-chart.component.scss
@@ -1,0 +1,7 @@
+
+#divChart{
+    display: block;
+    margin-top: 80px;
+    width: 100%;
+    height: 700px;
+}
