+++ src/app/cmps/contact-list/contact-list.component.scss
@@ -28,7 +28,7 @@
       }
 
       .icons {
-        color: black;
+        color: #457b9d;
         padding: 10px;
         opacity: 0;
