+++ src/app/pages/contact-details/contact-details.component.scss
@@ -1,21 +1,31 @@
 .contact-details {
-    display: flex;
-    flex-direction: column;
-    width: 450px;
-    font-size: 20px;
-    gap: 10px;
+
+    width: 500px;
     margin: 50px auto;
-    background-color: #dee2e6;
+    background-color: aliceblue;
     padding: 20px;
     border-radius: 8px;
-    color: black;
+    font-family: 'Outfit', sans-serif;
 
-    button {
-        border-radius: 8px;
-        font-size: 16px;
-        background-color: #495057;
-        width: 120px;
-        padding: 4px 0;
-        color: white;
+    .contact-info {
+        display: flex;
+        flex-direction: column;
+        font-size: 20px;
+        gap: 10px;
+        margin-top: 25px;
+
+        img{
+            border-radius: 50%;
+        }
+    }
+
+    .icon {
+        color: #457b9d;
+        font-size: 24px;
+        cursor: pointer;
+
+        &.edit {
+            float: right;
+        }
     }
 }
