+++ src/app/pages/stats/stats.component.html
@@ -1,6 +1,4 @@
 
 <section class="main-container">
-  <div id="divChart">
-    <canvas id="myChart" width="800"></canvas>
-  </div>
+  <avg-months-chart></avg-months-chart>
 </section>
