+++ src/app/pages/stats/stats.component.scss
@@ -2,10 +2,3 @@
 @import "../../../assets/styles/mixins";
 @import "../../../assets/styles/layout";
 
-
-#divChart{
-    display: block;
-    margin-top: 80px;
-    width: 100%;
-    height: 700px;
-}
